+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Unit tests for `MealService/getMeals`.
 - [Marek Kurańda](https://github.com/mjkuranda): E2E tests for `GET /meals`.
+- [Marek Kurańda](https://github.com/mjkuranda): Unit tests for `AbstractApiService/getMeals`.
+- [Marek Kurańda](https://github.com/mjkuranda): `AxiosService` for wrapping axios instance to inject to API integration services.
 
 ## [1.21.1] - 2024-02-12
 ### Fixed